+++ ui/components/app/loading-network-screen/loading-network-screen.component.js
@@ -1,9 +1,24 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import Button from '../../ui/button';
 import LoadingScreen from '../../ui/loading-screen';
 import { SECOND } from '../../../../shared/constants/time';
 import { NETWORK_TYPES } from '../../../../shared/constants/network';
+import Popover from '../../ui/popover/popover.component';
+import {
+  ButtonPrimary,
+  ButtonSecondary,
+  Icon,
+  IconName,
+  IconSize,
+  Text,
+} from '../../component-library';
+import {
+  DISPLAY,
+  IconColor,
+  TextAlign,
+  TextVariant,
+} from '../../../helpers/constants/design-system';
+import Box from '../../ui/box/box';
 
 export default class LoadingNetworkScreen extends PureComponent {
   state = {
@@ -17,7 +32,7 @@ export default class LoadingNetworkScreen extends PureComponent {
   static propTypes = {
     loadingMessage: PropTypes.string,
     cancelTime: PropTypes.number,
-    provider: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+    providerConfig: PropTypes.object,
     providerId: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
     showNetworkDropdown: PropTypes.func,
     setProviderArgs: PropTypes.array,
@@ -38,8 +53,8 @@ export default class LoadingNetworkScreen extends PureComponent {
     if (loadingMessage) {
       return loadingMessage;
     }
-    const { provider, providerId } = this.props;
-    const providerName = provider.type;
+    const { providerConfig, providerId } = this.props;
+    const providerName = providerConfig.type;
     const { t } = this.context;
 
     switch (providerName) {
@@ -56,63 +71,76 @@ export default class LoadingNetworkScreen extends PureComponent {
     }
   };
 
-  renderDeprecatedRpcUrlWarning = () => {
-    const { showNetworkDropdown } = this.props;
-
-    return (
-      <div className="loading-overlay__error-screen">
-        <span className="loading-overlay__emoji">&#128542;</span>
-        <span>{this.context.t('currentRpcUrlDeprecated')}</span>
-        <div className="loading-overlay__error-buttons">
-          <Button
-            type="secondary"
-            onClick={() => {
-              window.clearTimeout(this.cancelCallTimeout);
-              showNetworkDropdown();
-            }}
-          >
-            {this.context.t('switchNetworks')}
-          </Button>
-        </div>
-      </div>
-    );
-  };
-
-  renderErrorScreenContent = () => {
+  renderConnectionFailureNotification = (message, showTryAgain = false) => {
     const { showNetworkDropdown, setProviderArgs, setProviderType } =
       this.props;
 
     return (
-      <div className="loading-overlay__error-screen">
-        <span className="loading-overlay__emoji">&#128542;</span>
-        <span>{this.context.t('somethingWentWrong')}</span>
-        <div className="loading-overlay__error-buttons">
-          <Button
-            type="secondary"
+      <Popover
+        onClose={() => {
+          window.clearTimeout(this.cancelCallTimeout);
+        }}
+        centerTitle
+        title={
+          <Icon
+            name={IconName.Danger}
+            size={IconSize.Xl}
+            color={IconColor.warningDefault}
+          />
+        }
+      >
+        <Text
+          variant={TextVariant.bodyLgMedium}
+          textAlign={TextAlign.Center}
+          margin={[0, 4, 4, 4]}
+        >
+          {message}
+        </Text>
+        <Box display={DISPLAY.FLEX} padding={4} gap={2}>
+          <ButtonSecondary
             onClick={() => {
               window.clearTimeout(this.cancelCallTimeout);
               showNetworkDropdown();
             }}
+            variant={TextVariant.bodySm}
+            block
           >
             {this.context.t('switchNetworks')}
-          </Button>
+          </ButtonSecondary>
+          {showTryAgain ? (
+            <ButtonPrimary
+              onClick={() => {
+                this.setState({ showErrorScreen: false });
+                setProviderType(...setProviderArgs);
+                window.clearTimeout(this.cancelCallTimeout);
+                this.cancelCallTimeout = setTimeout(
+                  this.cancelCall,
+                  this.props.cancelTime || SECOND * 15,
+                );
+              }}
+              variant={TextVariant.bodySm}
+              block
+            >
+              {this.context.t('tryAgain')}
+            </ButtonPrimary>
+          ) : null}
+        </Box>
+      </Popover>
+    );
+  };
 
-          <Button
-            type="primary"
-            onClick={() => {
-              this.setState({ showErrorScreen: false });
-              setProviderType(...setProviderArgs);
-              window.clearTimeout(this.cancelCallTimeout);
-              this.cancelCallTimeout = setTimeout(
-                this.cancelCall,
-                this.props.cancelTime || SECOND * 15,
-              );
-            }}
-          >
-            {this.context.t('tryAgain')}
-          </Button>
-        </div>
-      </div>
+  renderDeprecatedRpcUrlWarning = () => {
+    return this.renderConnectionFailureNotification(
+      this.context.t('currentRpcUrlDeprecated'),
+      false,
+    );
+  };
+
+  renderErrorScreenContent = () => {
+    const { providerConfig } = this.props;
+    return this.renderConnectionFailureNotification(
+      this.context.t('networkSwitchConnectionError', [providerConfig.nickname]),
+      true,
     );
   };
 
@@ -125,9 +153,9 @@ export default class LoadingNetworkScreen extends PureComponent {
   };
 
   componentDidUpdate = (prevProps) => {
-    const { provider } = this.props;
-    const { provider: prevProvider } = prevProps;
-    if (provider.type !== prevProvider.type) {
+    const { providerConfig } = this.props;
+    const { providerConfig: prevProvider } = prevProps;
+    if (providerConfig.type !== prevProvider.type) {
       window.clearTimeout(this.cancelCallTimeout);
       this.setState({ showErrorScreen: false });
       this.cancelCallTimeout = setTimeout(