+++ ui/components/app/snaps/snap-delineator/snap-delineator.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DelineatorType } from '../../../../helpers/constants/flask';
+import { DelineatorType } from '../../../../helpers/constants/snaps';
 import { renderWithLocalization } from '../../../../../test/lib/render-helpers';
 import { SnapDelineator } from './snap-delineator';