+++ ui/pages/routes/routes.component.js
@@ -34,6 +34,7 @@ import {
   AppHeader as MultichainAppHeader,
   AccountListMenu,
   NetworkListMenu,
+  AccountDetails,
 } from '../../components/multichain';
 import UnlockPage from '../unlock-page';
 import Alerts from '../../components/app/alerts';
@@ -152,6 +153,7 @@ export default class Routes extends Component {
     toggleAccountMenu: PropTypes.func,
     isNetworkMenuOpen: PropTypes.bool,
     toggleNetworkMenu: PropTypes.func,
+    accountDetailsAddress: PropTypes.string,
   };
 
   static contextTypes = {
@@ -495,6 +497,8 @@ export default class Routes extends Component {
       toggleAccountMenu,
       isNetworkMenuOpen,
       toggleNetworkMenu,
+      accountDetailsAddress,
+      location,
     } = this.props;
     const loadMessage =
       loadingMessage || isNetworkLoading
@@ -539,7 +543,7 @@ export default class Routes extends Component {
         <Alert visible={this.props.alertOpen} msg={alertMessage} />
         {!this.hideAppHeader() &&
           (process.env.MULTICHAIN ? (
-            <MultichainAppHeader />
+            <MultichainAppHeader location={location} />
           ) : (
             <AppHeader
               hideNetworkIndicator={this.onInitializationUnlockPage()}
@@ -565,6 +569,9 @@ export default class Routes extends Component {
         ) : null}
         {process.env.MULTICHAIN && isNetworkMenuOpen ? (
           <NetworkListMenu onClose={() => toggleNetworkMenu()} />
+        ) : null}
+        {process.env.MULTICHAIN && accountDetailsAddress ? (
+          <AccountDetails address={accountDetailsAddress} />
         ) : null}
         <div className="main-container-wrapper">
           {isLoading ? <Loading loadingMessage={loadMessage} /> : null}